Extract LazyDiskCache to a separated file
